move branch-cache files to /cache
