+++ lib/util/cache/branch/behind-base-branch-cache.spec.ts
@@ -1,6 +1,6 @@
-import { logger, mocked, partial } from '../../../test/util';
-import * as _repositoryCache from '../cache/repository';
-import type { BranchCache, RepoCacheData } from '../cache/repository/types';
+import { logger, mocked, partial } from '../../../../test/util';
+import * as _repositoryCache from '../repository';
+import type { BranchCache, RepoCacheData } from '../repository/types';
 import {
   getCachedBehindBaseResult,
   setCachedBehindBaseResult,
@@ -9,7 +9,7 @@ import {
 jest.mock('../cache/repository');
 const repositoryCache = mocked(_repositoryCache);
 
-describe('util/git/behind-base-branch-cache', () => {
+describe('util/cache/branch/behind-base-branch-cache', () => {
   let repoCache: RepoCacheData = {};
 
   beforeEach(() => {
